refactor: zap transaction controls component
